pyproject.toml
[tool.poetry]
name = "azure-tools"
version = "0.1.0"
description = ""
authors = ["Matteo Campinoti <matteo.campinoti94@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
azure-storage-file-share = "^12.15.0"
azure-storage-blob = "^12.19.0"
click = "^8.1.7"

[tool.poetry.scripts]
"azure-tools" = "azure_tools.__main__:__main__"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
